refactor: Card 컴포넌트와 CreditCard 컴포넌트 분리
