+++ src/components/commons/card/Card.js
@@ -1,3 +1,4 @@
+import { CARD_COLOR } from '../../../constants/color';
 import Styled from './Card.style';
 
 export const CARD_SIZE = {
@@ -5,19 +6,10 @@ export const CARD_SIZE = {
   MD: 'MD',
 };
 
-export const CARD_COLOR = {
-  INDIGO: '#547CE4',
-  LIGHT_GRAY: '#D2D2D2',
-};
-
-export const Card = ({ size, color, content }) => {
+export const Card = ({ size, color, children, ...props }) => {
   return (
-    <Styled.Card size={size} color={color}>
-      <Styled.CardType size={size}>{content.cardType}</Styled.CardType>
-      <Styled.Chip size={size} />
-      <Styled.CardNumber size={size}>{content.cardNumber}</Styled.CardNumber>
-      <Styled.CardOwner size={size}>{content.cardOwner || 'NAME'}</Styled.CardOwner>
-      <Styled.CardExpiredDate size={size}>{content.cardExpiredDate || 'MM / YY'}</Styled.CardExpiredDate>
+    <Styled.Card size={size} color={color} {...props}>
+      {children}
     </Styled.Card>
   );
 };
@@ -25,8 +17,4 @@ export const Card = ({ size, color, content }) => {
 Card.defaultProps = {
   size: CARD_SIZE.MD,
   color: CARD_COLOR.LIGHT_GRAY,
-  content: {
-    cardOwner: 'NAME',
-    cardExpiredDate: 'MM / YY',
-  },
 };
