+++ src/stories/Card.stories.js
@@ -1,30 +1,28 @@
 import { CARD_SIZE, Card } from '../components/commons/card/Card';
+import { COLOR } from '../constants/color';
 
 export default {
   title: 'Card',
   component: Card,
+  argTypes: {
+    children: { control: 'text' },
+    backgroundColor: {
+      options: COLOR,
+      control: { type: 'select' },
+    },
+  },
 };
 
-const Template = (args) => <Card {...args} />;
+const Template = args => <Card {...args} />;
 
 export const Medium = Template.bind({});
 Medium.args = {
   size: CARD_SIZE.MD,
-  content: {
-    cardType: 'XX 카드',
-    cardNumber: '1111 2222 3333 4444',
-    cardOwner: 'Name',
-    cardExpiredDate: 'MM / YY',
-  },
+  color: COLOR.LIGHT_GRAY,
 };
 
 export const Large = Template.bind({});
 Large.args = {
   size: CARD_SIZE.LG,
-  content: {
-    cardType: 'XX 카드',
-    cardNumber: '1111 2222 3333 4444',
-    cardOwner: 'Name',
-    cardExpiredDate: 'MM / YY',
-  },
+  color: COLOR.LIGHT_GRAY,
 };
